Default config for denovo gene sets changed.
