+++ dae/dae/gene/tests/conftest.py
@@ -3,8 +3,11 @@
 import os
 import shutil
 
+import textwrap
 import pytest
 
+from dae.testing import setup_pedigree, setup_denovo, denovo_study
+from dae.testing.foobar_import import foobar_gpf
 
 from dae.utils.fixtures import change_environment
 
@@ -126,3 +129,44 @@ def denovo_gene_set_f4(local_gpf_instance):
 @pytest.fixture(scope="session")
 def f4_trio_denovo_gene_set_config(local_gpf_instance):
     return local_gpf_instance.get_genotype_data_config("f4_trio")
+
+
+@pytest.fixture
+def trios2_study(tmp_path_factory):
+    root_path = tmp_path_factory.mktemp(
+        "denovo_gene_sets_tios")
+    gpf_instance = foobar_gpf(root_path)
+    ped_path = setup_pedigree(
+        root_path / "trios2_data" / "in.ped",
+        """
+        familyId personId dadId	 momId	sex status role
+        f1       m1       0      0      2   1      mom
+        f1       d1       0      0      1   1      dad
+        f1       p1       d1     m1     1   2      prb
+        f2       m2       0      0      2   1      mom
+        f2       d2       0      0      1   1      dad
+        f2       p2       d2     m2     1   2      prb
+        """)
+    vcf_path = setup_denovo(
+        root_path / "trios2_data" / "in.tsv",
+        """
+          familyId  location  variant    bestState
+          f1        foo:10    sub(A->G)  2||2||1||1/0||0||1||0
+          f1        foo:11    sub(T->A)  2||2||1||2/0||0||1||0
+          f1        bar:10    sub(G->A)  2||2||2||1/0||0||0||1
+          f2        bar:11    sub(G->A)  2||2||1/0||0||1
+          f2        bar:12    sub(G->A)  2||2||1/0||0||1
+        """
+    )
+
+    study = denovo_study(
+        root_path,
+        "trios2", ped_path, [vcf_path],
+        gpf_instance,
+        study_config_update=textwrap.dedent("""
+        id: trios2
+        denovo_gene_sets:
+          enabled: True
+        """))
+    return study
+
