+++ dae/dae/gene/tests/test_denovo_gene_set.py
@@ -1,3 +1,4 @@
+# pylint: disable=W0621,C0114,C0116,W0212,W0613
 import pytest
 
 pytestmark = pytest.mark.usefixtures("gene_info_cache_dir", "calc_gene_sets")
