+++ dae/dae/gene/tests/test_denovo_gene_sets_config.py
@@ -1,50 +1,4 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
-import textwrap
-import pytest
-
-from dae.testing import setup_pedigree, setup_denovo, denovo_study
-from dae.testing.foobar_import import foobar_gpf
-
-
-@pytest.fixture
-def trios2_study(tmp_path_factory):
-    root_path = tmp_path_factory.mktemp(
-        "denovo_gene_sets_tios")
-    gpf_instance = foobar_gpf(root_path)
-    ped_path = setup_pedigree(
-        root_path / "trios2_data" / "in.ped",
-        """
-        familyId personId dadId	 momId	sex status role
-        f1       m1       0      0      2   1      mom
-        f1       d1       0      0      1   1      dad
-        f1       p1       d1     m1     1   2      prb
-        f2       m2       0      0      2   1      mom
-        f2       d2       0      0      1   1      dad
-        f2       p2       d2     m2     1   2      prb
-        """)
-    vcf_path = setup_denovo(
-        root_path / "trios2_data" / "in.tsv",
-        """
-          familyId  location  variant    bestState
-          f1        foo:10    sub(A->G)  2||2||1||1/0||0||1||0
-          f1        foo:11    sub(T->A)  2||2||1||2/0||0||1||0
-          f1        bar:10    sub(G->A)  2||2||2||1/0||0||0||1
-          f2        bar:11    sub(G->A)  2||2||1/0||0||1
-          f2        bar:12    sub(G->A)  2||2||1/0||0||1
-        """
-    )
-
-    study = denovo_study(
-        root_path,
-        "trios2", ped_path, [vcf_path],
-        gpf_instance,
-        study_config_update=textwrap.dedent("""
-        id: trios2
-        denovo_gene_sets:
-          enabled: True
-        """))
-    return study
-
 
 def test_trios2_study_simple(trios2_study):
     config = trios2_study.config
