+++ dae_conftests/dae_conftests/tests/fixtures/studies/Study1/denovo-cache-status.json
@@ -1,21 +1,13 @@
 {
-    "phenotype1": {
+    "affected": {
         "LGDs": {
             "Female": {},
             "Male": {},
             "Unspecified": {}
         },
         "Missense": {
-            "Female": {
-                "SAMD11": [
-                    "f1"
-                ]
-            },
-            "Male": {
-                "SAMD11": [
-                    "f2"
-                ]
-            },
+            "Female": {},
+            "Male": {},
             "Unspecified": {}
         },
         "Synonymous": {
@@ -26,7 +18,7 @@
             },
             "Male": {
                 "SAMD11": [
-                    "f2"
+                    "f6"
                 ]
             },
             "Unspecified": {}
@@ -44,11 +36,7 @@
             "Unspecified": {}
         },
         "Synonymous": {
-            "Female": {
-                "PLEKHN1": [
-                    "f2"
-                ]
-            },
+            "Female": {},
             "Male": {},
             "Unspecified": {}
         }
