+++ dae_conftests/dae_conftests/tests/fixtures/studies/Study2/denovo-cache-status.json
@@ -1,34 +1,18 @@
 {
-    "phenotype1": {
+    "affected": {
         "LGDs": {
             "Female": {},
             "Male": {},
             "Unspecified": {}
         },
         "Missense": {
-            "Female": {
-                "SAMD11": [
-                    "f1"
-                ]
-            },
-            "Male": {
-                "SAMD11": [
-                    "f2"
-                ]
-            },
+            "Female": {},
+            "Male": {},
             "Unspecified": {}
         },
         "Synonymous": {
-            "Female": {
-                "SAMD11": [
-                    "f1"
-                ]
-            },
-            "Male": {
-                "SAMD11": [
-                    "f2"
-                ]
-            },
+            "Female": {},
+            "Male": {},
             "Unspecified": {}
         }
     },
@@ -44,11 +28,7 @@
             "Unspecified": {}
         },
         "Synonymous": {
-            "Female": {
-                "PLEKHN1": [
-                    "f2"
-                ]
-            },
+            "Female": {},
             "Male": {},
             "Unspecified": {}
         }
