+++ dae_conftests/dae_conftests/tests/fixtures/studies/Study3/denovo-cache-status.json
@@ -1,34 +1,30 @@
 {
-    "phenotype1": {
+    "affected": {
         "LGDs": {
-            "Female": {},
-            "Male": {},
-            "Unspecified": {}
-        },
-        "Missense": {
             "Female": {
-                "SAMD11": [
-                    "f1"
-                ]
-            },
-            "Male": {
+                "PLEKHN1": [
+                    "f2",
+                    "f4"
+                ],
                 "SAMD11": [
                     "f2"
                 ]
             },
+            "Male": {},
+            "Unspecified": {}
+        },
+        "Missense": {
+            "Female": {},
+            "Male": {},
             "Unspecified": {}
         },
         "Synonymous": {
             "Female": {
                 "SAMD11": [
-                    "f1"
-                ]
-            },
-            "Male": {
-                "SAMD11": [
                     "f2"
                 ]
             },
+            "Male": {},
             "Unspecified": {}
         }
     },
@@ -45,8 +41,8 @@
         },
         "Synonymous": {
             "Female": {
-                "PLEKHN1": [
-                    "f2"
+                "SAMD11": [
+                    "f4"
                 ]
             },
             "Male": {},
