+++ wdae/wdae/enrichment_api/tests/conftest.py
@@ -1,3 +1,5 @@
+# pylint: disable=W0621,C0114,C0116,W0212,W0613
+
 import pytest
 
 from enrichment_api.enrichment_builder import EnrichmentBuilder
